Move api call to hook
